Add BPFRewriter element

+++ libs/click/unikraft/bpfclassifier.cc
@@ -1,13 +1,6 @@
-/*
- * bPFClassifier.{cc,hh} -- element which filters packets based on an ebpf program.
- */
-
 #include <click/config.h>
 #include <click/confparse.hh>
-#include <click/error.hh>
 #include <click/standard/scheduleinfo.hh>
-#include <stdio.h>
-#include <stdlib.h>
 
 #include "bpfclassifier.hh"
 
@@ -20,7 +13,7 @@ void BPFClassifier::push(int, Packet *p) {
     uk_pr_debug("BPFClassifier: Received packet\n");
 
     uk_rwlock_rlock(&_lock);
-    uint64_t ret = this->exec(p);
+    int ret = this->exec(p);
     uk_rwlock_runlock(&_lock);
 
     if (ret == -1) {
